+++ src/Twilio/Rest/FlexApi/V1/AssessmentsOptions.php
@@ -15,26 +15,25 @@ use Twilio\Values;
 /**
  * PLEASE NOTE that this class contains preview products that are subject to change. Use them with caution. If you currently do not have developer preview access, please contact help@twilio.com.
  */
-abstract class InsightsSegmentsOptions {
+abstract class AssessmentsOptions {
     /**
      * @param string $token The Token HTTP request header
-     * @return FetchInsightsSegmentsOptions Options builder
+     * @return CreateAssessmentsOptions Options builder
      */
-    public static function fetch(string $token = Values::NONE): FetchInsightsSegmentsOptions {
-        return new FetchInsightsSegmentsOptions($token);
+    public static function create(string $token = Values::NONE): CreateAssessmentsOptions {
+        return new CreateAssessmentsOptions($token);
     }
 
     /**
-     * @param string[] $reservationId List of reservation Ids
      * @param string $token The Token HTTP request header
-     * @return ReadInsightsSegmentsOptions Options builder
+     * @return UpdateAssessmentsOptions Options builder
      */
-    public static function read(array $reservationId = Values::ARRAY_NONE, string $token = Values::NONE): ReadInsightsSegmentsOptions {
-        return new ReadInsightsSegmentsOptions($reservationId, $token);
+    public static function update(string $token = Values::NONE): UpdateAssessmentsOptions {
+        return new UpdateAssessmentsOptions($token);
     }
 }
 
-class FetchInsightsSegmentsOptions extends Options {
+class CreateAssessmentsOptions extends Options {
     /**
      * @param string $token The Token HTTP request header
      */
@@ -60,29 +59,16 @@ class FetchInsightsSegmentsOptions extends Options {
      */
     public function __toString(): string {
         $options = \http_build_query(Values::of($this->options), '', ' ');
-        return '[Twilio.FlexApi.V1.FetchInsightsSegmentsOptions ' . $options . ']';
+        return '[Twilio.FlexApi.V1.CreateAssessmentsOptions ' . $options . ']';
     }
 }
 
-class ReadInsightsSegmentsOptions extends Options {
+class UpdateAssessmentsOptions extends Options {
     /**
-     * @param string[] $reservationId List of reservation Ids
      * @param string $token The Token HTTP request header
      */
-    public function __construct(array $reservationId = Values::ARRAY_NONE, string $token = Values::NONE) {
-        $this->options['reservationId'] = $reservationId;
+    public function __construct(string $token = Values::NONE) {
         $this->options['token'] = $token;
-    }
-
-    /**
-     * The list of reservation Ids
-     *
-     * @param string[] $reservationId List of reservation Ids
-     * @return $this Fluent Builder
-     */
-    public function setReservationId(array $reservationId): self {
-        $this->options['reservationId'] = $reservationId;
-        return $this;
     }
 
     /**
@@ -103,6 +89,6 @@ class ReadInsightsSegmentsOptions extends Options {
      */
     public function __toString(): string {
         $options = \http_build_query(Values::of($this->options), '', ' ');
-        return '[Twilio.FlexApi.V1.ReadInsightsSegmentsOptions ' . $options . ']';
+        return '[Twilio.FlexApi.V1.UpdateAssessmentsOptions ' . $options . ']';
     }
 }