+++ src/Twilio/Rest/FlexApi/V1/AssessmentsPage.php
@@ -13,6 +13,9 @@ use Twilio\Http\Response;
 use Twilio\Page;
 use Twilio\Version;
 
+/**
+ * PLEASE NOTE that this class contains preview products that are subject to change. Use them with caution. If you currently do not have developer preview access, please contact help@twilio.com.
+ */
 class AssessmentsPage extends Page {
     /**
      * @param Version $version Version that contains the resource