+++ src/Twilio/Rest/Microvisor/V1/AccountSecretInstance.php
@@ -68,6 +68,17 @@ class AccountSecretInstance extends InstanceResource {
     }
 
     /**
+     * Update the AccountSecretInstance
+     *
+     * @param string $value The secret value.
+     * @return AccountSecretInstance Updated AccountSecretInstance
+     * @throws TwilioException When an HTTP error occurs.
+     */
+    public function update(string $value): AccountSecretInstance {
+        return $this->proxy()->update($value);
+    }
+
+    /**
      * Delete the AccountSecretInstance
      *
      * @return bool True if delete succeeds, false otherwise