+++ src/Twilio/Rest/Taskrouter/V1/WorkspaceContext.php
@@ -17,7 +17,6 @@ use Twilio\Rest\Taskrouter\V1\Workspace\ActivityList;
 use Twilio\Rest\Taskrouter\V1\Workspace\EventList;
 use Twilio\Rest\Taskrouter\V1\Workspace\TaskChannelList;
 use Twilio\Rest\Taskrouter\V1\Workspace\TaskList;
-use Twilio\Rest\Taskrouter\V1\Workspace\TaskQueueBulkRealTimeStatisticsList;
 use Twilio\Rest\Taskrouter\V1\Workspace\TaskQueueList;
 use Twilio\Rest\Taskrouter\V1\Workspace\WorkerList;
 use Twilio\Rest\Taskrouter\V1\Workspace\WorkflowList;
@@ -39,7 +38,6 @@ use Twilio\Version;
  * @property WorkspaceRealTimeStatisticsList $realTimeStatistics
  * @property WorkspaceCumulativeStatisticsList $cumulativeStatistics
  * @property TaskChannelList $taskChannels
- * @property TaskQueueBulkRealTimeStatisticsList $bulkRealTimeStatistics
  * @method \Twilio\Rest\Taskrouter\V1\Workspace\ActivityContext activities(string $sid)
  * @method \Twilio\Rest\Taskrouter\V1\Workspace\EventContext events(string $sid)
  * @method \Twilio\Rest\Taskrouter\V1\Workspace\TaskContext tasks(string $sid)
@@ -62,7 +60,6 @@ class WorkspaceContext extends InstanceContext {
     protected $_realTimeStatistics;
     protected $_cumulativeStatistics;
     protected $_taskChannels;
-    protected $_bulkRealTimeStatistics;
 
     /**
      * Initialize the WorkspaceContext
@@ -240,20 +237,6 @@ class WorkspaceContext extends InstanceContext {
         }
 
         return $this->_taskChannels;
-    }
-
-    /**
-     * Access the bulkRealTimeStatistics
-     */
-    protected function getBulkRealTimeStatistics(): TaskQueueBulkRealTimeStatisticsList {
-        if (!$this->_bulkRealTimeStatistics) {
-            $this->_bulkRealTimeStatistics = new TaskQueueBulkRealTimeStatisticsList(
-                $this->version,
-                $this->solution['sid']
-            );
-        }
-
-        return $this->_bulkRealTimeStatistics;
     }
 
     /**