+++ tests/Twilio/Integration/Microvisor/V1/AccountConfigTest.php
@@ -159,6 +159,42 @@ class AccountConfigTest extends HolodeckTestCase {
         $this->assertNotNull($actual);
     }
 
+    public function testUpdateRequest(): void {
+        $this->holodeck->mock(new Response(500, ''));
+
+        try {
+            $this->twilio->microvisor->v1->accountConfigs("key")->update("value");
+        } catch (DeserializeException $e) {}
+          catch (TwilioException $e) {}
+
+        $values = ['Value' => "value", ];
+
+        $this->assertRequest(new Request(
+            'post',
+            'https://microvisor.twilio.com/v1/Configs/key',
+            null,
+            $values
+        ));
+    }
+
+    public function testUpdateResponse(): void {
+        $this->holodeck->mock(new Response(
+            200,
+            '
+            {
+                "key": "first",
+                "value": "place",
+                "date_updated": "2021-01-01T12:34:56Z",
+                "url": "https://microvisor.twilio.com/v1/Configs/first"
+            }
+            '
+        ));
+
+        $actual = $this->twilio->microvisor->v1->accountConfigs("key")->update("value");
+
+        $this->assertNotNull($actual);
+    }
+
     public function testDeleteRequest(): void {
         $this->holodeck->mock(new Response(500, ''));